requirements.txt
fastapi
uvicorn
piexif
Pillow